Adding test cases for groups.py file (#131)
* Adding test cases for groups.py file

* Removing extra else

* Removing extra spaces

* Adding test cases for group.py

* Fixed error handler for leave command

It doesn't have an argument.

+++ cogs/groups.py
@@ -38,6 +38,8 @@ class Groups(commands.Cog):
         for i in range(100):
             role_name = "group_" + str(i)
             role = get(ctx.message.guild.roles, name=role_name)
+            if role is None:
+                continue
             await role.delete()
 
         await ctx.author.send("Roles deleted!")
@@ -251,7 +253,7 @@ class Groups(commands.Cog):
 
     # -------------------------------------------------------------------------------------------------------
     #    Function: leave_error(self, ctx, error)
-    #    Description: prints error message for join command
+    #    Description: prints error message for leave command
     #    Inputs:
     #       - ctx: context of the command
     #       - error: error message
@@ -260,12 +262,9 @@ class Groups(commands.Cog):
     # -------------------------------------------------------------------------------------------------------
     @leave.error
     async def leave_error(self, ctx, error):
-        if isinstance(error, commands.MissingRequiredArgument):
-            await ctx.send('To use the join command, do: $join <Num> \n ( For example: $join 0 )')
-        else:
-            await ctx.author.send(error)
-            #await ctx.message.delete()
-            print(error)
+        await ctx.author.send(error)
+        #await ctx.message.delete()
+        print(error)
 
 
     # -------------------------------------------------------------------------------------------------------
